Fix the issue of avoiding app crashes in HWUI (Hardware UI)
During monkey testing, there is a probability that the Paint shader is released twice, leading to a crash. However, this issue cannot be reproduced through manual operations at present.
Through code analysis, the core of the problem is that when assigning the shader, an unref is triggered once, and when the Java object is released, another release is triggered, resulting in a double free.

Bug: 439735342
Bug: 431645587

Test: pressure test

Change-Id: I954eb739e39ac45106186424cb437f264ef10215