[res] Avoid appending shared libraries to outdated Assets
App A can create resources for app B through methods such as getResourcesForApplication.
These resources will be cached by the ResourceManager in application A's process.
If the package of application B changes, the cached resources in process A will become invalid.
If process A uses appendLibAssetsLocked to attach shared libraries to these outdated resources,
an exception will occur during loadApkAssets, causing application A to crash.

Therefore, when appendLibAssetsLocked, we need to add a check
to ensure that the ResourcesKey used to create ResourceImpl
does not contain an invalid apk path.

Bug:
Test:

Change-Id: Ie927a495fff5b2a23430b9076d633a25f8ba805b
Signed-off-by: zhangshuai10 <zhangshuai10@xiaomi.corp-partner.google.com>