[Audio] clear mute state before set new port ID
Reason:
The `mutedState` in `AudioPlaybackConfiguration` can become stale when PORT ID changed.
The mute state from the previous port is not cleared.

Solution:
when a `PLAYER_UPDATE_PORT_ID` event is received, check if an old port ID existed.
If so, dispatch a  `MSG_IIL_UPDATE_PLAYER_MUTED_EVENT` with a mute reason of 0 to
clear the mute state before the new port ID is assigned.

Test:repor steps in bug
Bug:444376896
Change-Id: I0ecc82b77ae948e3e31ea2bce69daf7459556ed2