Reduce the number of calculations in the dirty area.
When mHaveNewSurface is true, or mLastFrameWidth is 0, or mLastFrameHeight
is 0, it makes no sense to calculate the dirty area using DamageAccumulator,
because in CanvasContext::computeDirtyRect, the dirty area will be set to
be synchronized with the width and height of the Frame.

So when the above conditions are met, the Dirty Area will not be calculated
using the DamageAccumulator.

Change-Id: I87e062cfb8926aee351d33742219dd6baeeb7aae
Signed-off-by: Liangliang Sui <coolsui.coding@gmail.com>