Colemak keyboard layout: Keep Caps Lock as default
This lines up with the implementation of Colemak in Windows 11, macOS
and iOS, and the default Colemak installer for Windows 10.
The Caps Lock indicator was previously toggled on keyboards when
pressing the Caps Lock key, even though Caps Lock itself wasn't toggled.
Removing the Caps->Backspace mod is simpler for most users, and is
similar to how the Workman keyboard layout is implemented in AOSP.

Change-Id: I6997c3a84f5f21dc2ddf0a8f192c7e5834ae990b
Signed-off-by: Callum Andrew <dev@candrew.net>