Move broadcastReceiver to a singleton class
MobileConnectionRepositoryImpl is not a singleton. When
users repeatedly enable and disable Bluetooth connections,
multiple instances may be created, resulting in repeated
registration of the broadcast receiver.
However, since the scope's lifecycle is consistent with
the application's lifecycle, the awaitClose logic for
unregistering the receiver is never triggered, which can
lead to memory leaks.
Moving the broadcast receiver to the singleton class
MobileConnectionsRepositoryImpl can prevent the issue
of repeated registrations.

Bug: 458164217
Test: Manual
1.Repeated connect-disconnect-connect operation in Setting with a Bluetooth device which has a sim card

Change-Id: I6e518d06503180ca628e508effe26edd71e14538