Fix Activity removing from StoppingActivities but not resumed
Fast switching between two apps: When the onPause call of App B is slow, App A has already been removed from the StoppingActivities list. At this time, App B calls resume again, causing App A to not truly resume and get stuck in the onPause state—even though the app is already in the background.

Bug: 458567546
Change-Id: I40cf9cfa76dc40b61e709c7f8fd2206abce29a0c