System_server killed by watchdog: Blocked in handler on on main thread
system server will be killed while waiting for boot user to be set from car service by watchdog since systemserver blocks its main thread for 5min for the bootuser to be set, but as per design system server main thread can be stuck only for 60s.
This change is to reduce the BOOT_USER_SET_TIMEOUT_MS from 5min to 50s, to avoid this system_server watchdog kill due to getbootuser since there are some conditions that can be checked if boot user failed to be set.

Bug: 463261053
Test: Fleet testing

Signed-off-by: Arjun Sankaran <asanka17@ford.com>
Change-Id: I9b4f2f10d04a039684ad8fe29456e4a4d0ef6358